Update tests to yeoman-test@7.3.0 and mjs
